Rename: innerSection으로 파일명 수정 및 경로 이동
